PIM-7132: Fix "url too long" when doing a mass edit
Context:
When doing a mass-edit in 1.6 and 1.7 (mass-edit has been reworked in
2.0 <3 @juliensnz), the ids of the products selected in the
datagrid would be passed on all the mass-edit steps through the
URL parameter "values".

When submitting the form to create the mass-edit operation, the form
would be submitted via a POST http verb, but the selected Ids would still be
passed on to the controller using the URI parameter.

This submit action would sometimes not execute because the URL would be
too long.

Solution:
*1st part*: Passing the selected Ids from the datagrid to the mass-edit
(without using a url query parameter)

First thing, we need to get rid of the mecanism that passes on the
selected ids to the mass-edit using the URL.

To do so, we will serialize the selected ids in the browser's
localstore prior to go on the mass-edit configuration pages.

Those Ids will stay in the localstorage until the "Configure" step: at
this stage, we will put the Ids in a hidden input in the configure form
to be able to send it to the back-end using form-data.

First part done.

*2nd part*: Sending the selected item ids via POST form data
Instead of passing on the selected ids through an url parameter, we used
a hidden input (see above) in the mass-edit form.

When submitting the form to run the mass-edit, we now need to fetch
the product ids in the right form-data field: this work is mainly done in the
'Pim\Bundle\DataGridBundle\Extension\MassAction\MassActionDispatcher::prepareMassActionParameters`

As this class is also used for the Quick-exports, we need to keep the
old way of getting the value (through a URL parameter) in order to not
break the quick-exports for now.

Fix complete.